replace alias_method_chain
