+++ config/initializers/10-patches.rb
@@ -41,90 +41,6 @@ module ActiveRecord
   end
 end
 
-module ActiveModel
-  class Errors
-    ##
-    # ActiveRecord errors do provide no means to access the symbols initially used to create an
-    # error. E.g. errors.add :foo, :bar instantly translates :bar, making it hard to write code
-    # dependent on specific errors (which we use in the APIv3).
-    # We therefore add a second information store containing pairs of [symbol, translated_message].
-    def add_with_storing_error_symbols(attribute, message = :invalid, options = {})
-      error_symbol = options.fetch(:error_symbol) { message }
-      add_without_storing_error_symbols(attribute, message, options)
-
-      if store_new_symbols?
-        if error_symbol.is_a?(Symbol)
-          symbol = error_symbol
-          partial_message = normalize_message(attribute, message, options)
-          full_message = full_message(attribute, partial_message)
-        else
-          symbol = :unknown
-          full_message = message
-        end
-
-        writable_symbols_and_messages_for(attribute) << [symbol, full_message, partial_message]
-      end
-    end
-
-    alias_method_chain :add, :storing_error_symbols
-
-    def symbols_and_messages_for(attribute)
-      writable_symbols_and_messages_for(attribute).dup
-    end
-
-    def symbols_for(attribute)
-      symbols_and_messages_for(attribute).map(&:first)
-    end
-
-    def full_message(attribute, message)
-      return message if attribute == :base
-
-      # if a model acts_as_customizable it will inject attributes like 'custom_field_1' into itself
-      # using attr_name_override we resolve names of such attributes.
-      # The rest of the method should reflect the original method implementation of ActiveModel
-      attr_name_override = nil
-      match = /\Acustom_field_(?<id>\d+)\z/.match(attribute)
-      if match
-        attr_name_override = CustomField.find_by(id: match[:id]).name
-      end
-
-      attr_name = attribute.to_s.gsub('.', '_').humanize
-      attr_name = @base.class.human_attribute_name(attribute, default: attr_name)
-      I18n.t(:"errors.format",                                default: '%{attribute} %{message}',
-                                                              attribute: attr_name_override || attr_name,
-                                                              message: message)
-    end
-
-    # Need to do the house keeping along with AR::Errors
-    # so that the symbols are removed when a new validation round starts
-    def clear_with_storing_error_symbols
-      clear_without_storing_error_symbols
-
-      @error_symbols = Hash.new
-    end
-
-    alias_method_chain :clear, :storing_error_symbols
-
-    private
-
-    def error_symbols
-      @error_symbols ||= Hash.new
-    end
-
-    def writable_symbols_and_messages_for(attribute)
-      error_symbols[attribute.to_sym] ||= []
-    end
-
-    # Kind of a hack: We need the possibility to temporarily disable symbol storing in the subclass
-    # Reform::Contract::Errors, because otherwise we end up with duplicate entries
-    # I feel dirty for doing that, but on the other hand I see no other way out... Please, stop me!
-    def store_new_symbols?
-      @store_new_symbols = true if @store_new_symbols.nil?
-      @store_new_symbols
-    end
-  end
-end
-
 module ActionView
   module Helpers
     module Tags
@@ -218,14 +134,6 @@ module ActionView
           end
         end
       end
-    end
-
-    module AssetTagHelper
-      def auto_discovery_link_tag_with_no_atom_feeds(type = :rss, url_options = {}, tag_options = {})
-        return if (type == :atom) && Setting.table_exists? && !Setting.feeds_enabled?
-        auto_discovery_link_tag_without_no_atom_feeds(type, url_options, tag_options)
-      end
-      alias_method_chain :auto_discovery_link_tag, :no_atom_feeds
     end
   end
 end
